Update go-for-it to 1.6.3
Because all active development has been taken over by JMoerman,
this update also bring the issue of a new application id.

Since flatpak, and by extension flathub, cannot handle this change
gracefully a workaround is needed. Until there is a way of handling this
the go-for-it id is patched to remain `de.manuel_kehl.go-for-it`.

Additionally, as a result of upstream collaboration this update include
official AppStream appdata.